speedController: fixed overflow caused by unsigned in calculation

+++ firmware/common/speed_controller.c
@@ -76,7 +76,7 @@ int32_t speedControllerStep(struct ControllerTargetData *targetData, int32_t whe
     int32_t targetSpeed = *speed_p;
     
     //input is turns per second * 100
-    targetSpeed = targetSpeed * ticksPerTurn / (1000 * 100);
+    targetSpeed = targetSpeed * ((int32_t) ticksPerTurn) / (1000 * 100);
 
     int32_t pwmValue = 0;
